+++ internal/metrics/tags.go
@@ -9,5 +9,6 @@ var (
 	keyStatus      tag.Key = tag.MustNewKey("status")
 	keyService     tag.Key = tag.MustNewKey("service")
 	keyGRPCService tag.Key = tag.MustNewKey("grpc_service")
+	keyGRPCMethod  tag.Key = tag.MustNewKey("grpc_method")
 	keyHost        tag.Key = tag.MustNewKey("host")
 )