+++ internal/metrics/view_test.go
@@ -6,29 +6,20 @@ import (
 	"go.opencensus.io/stats/view"
 )
 
-func Test_RegisterHTTPClientView(t *testing.T) {
-	RegisterHTTPClientView()
-	for _, v := range []*view.View{HTTPClientRequestCountView, HTTPClientRequestDurationView, HTTPClientResponseSizeView} {
+func Test_RegisterView(t *testing.T) {
+	RegisterView(HTTPClientViews)
+	for _, v := range HTTPClientViews {
 		if view.Find(v.Name) != v {
 			t.Errorf("Failed to find registered view %s", v.Name)
 		}
 	}
 }
 
-func Test_RegisterHTTPServerView(t *testing.T) {
-	RegisterHTTPServerView()
-	for _, v := range []*view.View{HTTPServerRequestCountView, HTTPServerRequestDurationView, HTTPServerRequestSizeView} {
-		if view.Find(v.Name) != v {
-			t.Errorf("Failed to find registered view %s", v.Name)
-		}
-	}
-}
-
-func Test_RegisterGRPCClientView(t *testing.T) {
-	RegisterGRPCClientView()
-	for _, v := range []*view.View{GRPCClientRequestCountView, GRPCClientRequestDurationView, GRPCClientResponseSizeView} {
-		if view.Find(v.Name) != v {
-			t.Errorf("Failed to find registered view %s", v.Name)
+func Test_UnregisterView(t *testing.T) {
+	UnRegisterView(HTTPClientViews)
+	for _, v := range HTTPClientViews {
+		if view.Find(v.Name) == v {
+			t.Errorf("Found unregistered view %s", v.Name)
 		}
 	}
 }